Modified aws policy sample to match actual policy
The policy required to read SQS includes some policy actions that are not present in the policy sample. Actually the policy sample looks like an example to push data to SQS. To clarify the documentation I suggest to add the required policy actions of this plugin to the policy sample.